3rd commit: project completed

+++ src/main/java/calculator/Calculator.java
@@ -1,6 +1,19 @@
 package calculator;
 
+/***
+ * 
+ * @author Mudasir Habib
+ *
+ */
 public class Calculator {
+	
+	/***
+	 * findMax method find the maximumm number from 3 numbers that are passed in arguments
+	 * @param first number: num1 having type integer
+	 * @param second number: num2 having type integer
+	 * @param third number: num3 having type integer
+	 * @return it'll return maximum of three numbers
+	 */
 	public int findMax(int num1, int num2, int num3) {
 		int[] arr = { num1, num2, num3 };   
 		int max = 0;   
@@ -11,6 +24,13 @@ public class Calculator {
 		return max;
 		}  
 	
+	
+	/**
+	 * Square method calculates the square of a integer , which is greater than 0 and less than 10
+	 * @param integer type parameter
+	 * @return it'll return integer type of square of number
+	 * @throws IllegalArgumentException when the input integer is not between 0 and 10
+	 */	
 	public int square(int num) throws IllegalArgumentException {
 		int result = 0;   
 		if (num > 0 && num < 10) {
@@ -22,8 +42,14 @@ public class Calculator {
 		return result;  
 		} 
 	
-	public long cube(long number) throws IllegalArgumentException {
-		long cubeResult = 0;
+	/**
+	 * Cube method calculates the cube of a int type , which is greater than 0
+	 * @param int type parameter
+	 * @return it'll return int type of cube of number
+	 * @throws IllegalArgumentException when the input integer is not greater than 0
+	 */
+	public int cube(int number) throws IllegalArgumentException {
+		int cubeResult = 0;
 		if (number > 0) {
 			cubeResult = number * number * number;
 		}
